move files and separate bidir in module xml
