+++ conf/modules/actuators_dshot.xml
@@ -27,17 +27,19 @@
   </header>
   <makefile>
     <configure name="DSHOT_BIDIR" default="FALSE"/>
-    <define name="DSHOT_BIDIR" value="$(DSHOT_BIDIR)"/>
-    <define name="USE_GPT$(DSHOT1_GPT_TIM)" value="TRUE" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
-    <define name="DSHOT1_BIDIR_GPT" value="GPTD$(DSHOT1_GPT_TIM)" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
-
     <file_arch name="actuators_dshot_arch.c"/>
     <file_arch name="esc_dshot.c" cond="ifeq ($(RTOS),chibios)"/>
-    <file_arch name="dshot_rpmCapture.c" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
-    <file_arch name="dshot_erps.c" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
-    <file_arch name="input_capture_arch.c" dir="mcu_periph" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
-    <file_arch name="timerDmaCache.c" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
     <file_arch name="hal_stm32_dma.c" dir="mcu_periph" cond="ifeq ($(RTOS),chibios)"/>
+    <test/>
+  </makefile>
+  <makefile cond="ifeq ($(DSHOT_BIDIR),TRUE)">
+    <define name="DSHOT_BIDIR" value="$(DSHOT_BIDIR)"/>
+    <define name="USE_GPT$(DSHOT1_GPT_TIM)" value="TRUE"/>
+    <define name="DSHOT1_BIDIR_GPT" value="GPTD$(DSHOT1_GPT_TIM)"/>
+    <file_arch name="dshot_rpmCapture.c"/>
+    <file_arch name="dshot_erps.c"/>
+    <file_arch name="input_capture_arch.c" dir="mcu_periph"/>
+    <file_arch name="timerDmaCache.c" dir="mcu_periph"/>
     <test/>
   </makefile>
 </module>
